Put the right code in the right file
